refactor: split exporter tests and add file operations tests
- Extract test helpers into exporter_test_helper.go for better reusability
- Move export processor tests to dedicated file export_processor_test.go
- Add comprehensive file operations tests in file_operations_test.go
- Add stats test suite in stats_test.go
